Fixed Bayesian regression implementation in SBI workflows
