+++ .gitignore
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 .ipynb_checkpoints
 __pycache__
 *.swp
@@ -17,3 +18,6 @@ build/
 _build/
 dist/
 *.egg-info/
+=======
+"" 
+>>>>>>> 5d8ded8c (Added .gitignore file)
